\name{get_dataset_path}
\alias{get_dataset_path}
\title{Get file path to a packaged dataset}
\usage{
get_dataset_path(name)
}
\arguments{
\item{name}{Name of dataset CSV file (e.g. "course_dataset.csv").}
}
\value{
Absolute file path to the dataset.
}
\description{
Get file path to a packaged dataset in scnpir.
}
\examples{
\dontrun{
get_dataset_path("course_dataset.csv")
}
}
