\name{run_unit2}
\alias{run_unit2}
\title{Run Unit 2 Interactive Tutorial}
\usage{
run_unit2()
}
\value{
Launches the interactive tutorial in a web browser.
}
\description{
Run Unit 2 Interactive Tutorial (Descriptive Statistics and Programming Fundamentals in R).
}
\examples{
\dontrun{
run_unit2()
}
}
