\name{run_unit4}
\alias{run_unit4}
\title{Run Unit 4 Interactive Tutorial}
\usage{
run_unit4()
}
\value{
Launches the interactive tutorial in a web browser.
}
\description{
Run Unit 4 Interactive Tutorial (Analysis of Contingency Tables and Compact Reporting).
}
\examples{
\dontrun{
run_unit4()
}
}
